% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diff.argList}
\alias{diff.argList}
\title{diff}
\usage{
\method{diff}{argList}(x, y, xtitle = NULL, ytitle = NULL, ...)
}
\arguments{
\item{x, y}{argList: the list to compare}

\item{xtitle, ytitle}{character: titles for the argLists (default: \code{NULL} = \code{deparse1(substitute(.))})}

\item{...}{unused}
}
\value{
a character vector
}
\description{
Shows the differences and similarities between two argument lists in the style of Unix diff:
\itemize{
\item \verb{< ...} means that \code{...} is only \code{x},
\item \verb{> ...} means that \code{...} is only \code{y}, and
\item '  ...\verb{means that}...\verb{is in}x\code{and}y`.
}
}
\examples{
1+1
}
