% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argList.R
\name{update.argList}
\alias{update.argList}
\title{update}
\usage{
\method{update}{argList}(object, ...)
}
\arguments{
\item{object}{argument list: list of arguments}

\item{...}{named parameter for managing global and local parameters}
}
\value{
a modified argument list
}
\description{
Updates a argument list with global and local parameters
}
\examples{
1+1
}
